Import deterministic RSA key generation code from Go 1.10 crypto/rsa package
Go 1.11 introduced some explicit randomisation for some crypto routines, because
they don't want callers to rely on the fact that current algorithms are
deterministic with respect to the passed in pseudo-random stream (which is
exactly what this package does) to be able to change the underlying algorithms
in the future.

Since we do need the key generation algorithms to be deterministic, we just
import the algorithm to our package. This will also ensure consistent output
for generated RSA keys should the standard crypto/rsa package change its
generation algorithm.