Add a test-case which tracks changes to the stdlib RSA key generation algorithm
For now, we would like to track the stdlib RSA key generation algorithm, so
we add a test, which disables the current randomisation in the stdlib
implementation and compares its output to a known value.

We might remove this test in the future, if the stdlib algorithm will change
in a non backward compatible/backportable way.